backend/requirements.txt
asgiref==3.10.0
asttokens==3.0.0
black==25.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
decorator==5.2.1
defusedxml==0.7.1
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
executing==2.2.1
gunicorn==23.0.0
httpie==3.2.4
idna==3.11
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
markdown-it-py==4.0.0
matplotlib-inline==0.2.1
mdurl==0.1.2
mongoengine==0.29.1
multidict==6.7.0
mypy_extensions==1.1.0
packaging==25.0
parso==0.8.5
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.15.3
PySocks==1.7.1
python-dotenv==1.2.1
pytokens==0.2.0
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
setuptools==80.9.0
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
urllib3==2.5.0
wcwidth==0.2.14
